+++ app/src/main/java/com/adammcneilly/pocketleague/eventoverview/ui/EventOverviewViewModel.kt
@@ -3,9 +3,9 @@ package com.adammcneilly.pocketleague.eventoverview.ui
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.adammcneilly.pocketleague.shared.eventoverview.EventOverviewAction
 import com.adammcneilly.pocketleague.shared.eventoverview.domain.GetEventOverviewUseCase
-import com.adammcneilly.pocketleague.shared.eventoverview.eventOverviewStateMutator
+import com.adammcneilly.pocketleague.shared.eventoverview.state.EventOverviewAction
+import com.adammcneilly.pocketleague.shared.eventoverview.state.eventOverviewStateMutator
 import com.ramcosta.composedestinations.EventOverviewScreenDestination
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
