+++ shared/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/eventsummarylist/state/EventSummaryListViewState.kt
@@ -1,6 +1,8 @@
-package com.adammcneilly.pocketleague.shared.eventsummarylist
+package com.adammcneilly.pocketleague.shared.eventsummarylist.state
 
 import com.adammcneilly.pocketleague.shared.core.ui.UIText
+import com.adammcneilly.pocketleague.shared.eventsummarylist.ui.EventSummaryListItemDisplayModel
+import com.adammcneilly.pocketleague.shared.eventsummarylist.EventSummaryListSort
 
 /**
  * A collection of possible view states for [EventSummaryListScreen].
